Refactor mixing, need to add tests
