+++ apps/covid_19/model/model.py
@@ -134,8 +134,12 @@ def build_model(params: dict) -> StratifiedModel:
     """
     Dynamic heterogeneous mixing by age
     """
-    static_mixing_matrix = preprocess.mixing_matrix.build_static(country.iso3)
-    dynamic_mixing_matrix = preprocess.mixing_matrix.build_dynamic(country, params.mobility)
+    static_mixing_matrix = inputs.get_country_mixing_matrix(country.iso3)
+    dynamic_mixing_matrix = preprocess.mixing_matrix.build_dynamic_mixing_matrix(
+        static_mixing_matrix,
+        params.mobility,
+        country,
+    )
     model.set_dynamic_mixing_matrix(dynamic_mixing_matrix)
 
     """
@@ -613,10 +617,12 @@ def build_model(params: dict) -> StratifiedModel:
         func_outputs["accum_years_of_life_lost"] = outputs.calculate_cum_years_of_life_lost
         for agegroup in agegroup_strata:
             age_key = f"agegroup_{agegroup}"
-            func_outputs[f"proportion_seropositiveX{age_key}"] = outputs.make_age_specific_seroprevalence_output(
+            func_outputs[
+                f"proportion_seropositiveX{age_key}"
+            ] = outputs.make_age_specific_seroprevalence_output(agegroup)
+            func_outputs[f"accum_deathsX{age_key}"] = outputs.make_agespecific_cum_deaths_func(
                 agegroup
             )
-            func_outputs[f"accum_deathsX{age_key}"] = outputs.make_agespecific_cum_deaths_func(agegroup)
 
     model.add_function_derived_outputs(func_outputs)
     return model
