+++ apps/covid_19/model/preprocess/mixing_matrix/__init__.py
@@ -1,3 +1,1 @@
-from .dynamic import build_dynamic
-from .static import build_static
-from .utils import get_total_contact_rates_by_age
+from .mixing_matrix import build_dynamic_mixing_matrix
